+++ src/App.js
@@ -1,13 +1,11 @@
 import "./App.css";
 import CardRows from "./components/cardRows";
-import buildDeck from "./utils/buildDeck";
-import SelectCards from "./utils/selectCards";
-import shuffleDeck from "./utils/shuffleDeck";
+import { buildDeck, shuffleDeck, selectCards } from "./utils/deck";
 
 function App() {
   const DECK = buildDeck();
   const SHUFFLED_DECK = shuffleDeck(DECK);
-  const SELECTED_CARDS = SelectCards(SHUFFLED_DECK);
+  const SELECTED_CARDS = selectCards(SHUFFLED_DECK);
 
   return (
     <div className="App">
